refactor: update package name to metamodel to follow the Jakarta Persistence nameclature
