+++ api/src/main/java/jakarta/data/metamodel/package-info.java
@@ -55,4 +55,4 @@
  * page1 = products.findByNameLike(namePattern, pageRequest);
  * </pre>
  */
-package jakarta.data.model;
+package jakarta.data.metamodel;
