+++ tck/src/main/resources/ee/jakarta/tck/data/framework/signature/jakarta.data.sig_17
@@ -93,32 +93,32 @@ hfds serialVersionUID
 
 CLSS abstract interface jakarta.data.exceptions.package-info
 
-CLSS public jakarta.data.model.Attribute
-meth public final boolean init(jakarta.data.model.AttributeInfo)
+CLSS public jakarta.data.metamodel.Attribute
+meth public final boolean init(jakarta.data.metamodel.AttributeInfo)
 meth public final jakarta.data.Sort asc()
 meth public final jakarta.data.Sort ascIgnoreCase()
 meth public final jakarta.data.Sort desc()
 meth public final jakarta.data.Sort descIgnoreCase()
 meth public final java.lang.String name()
-meth public final static jakarta.data.model.Attribute get()
+meth public final static jakarta.data.metamodel.Attribute get()
 supr java.lang.Object
 hfds info
 
-CLSS public abstract interface jakarta.data.model.AttributeInfo
+CLSS public abstract interface jakarta.data.metamodel.AttributeInfo
 meth public abstract jakarta.data.Sort asc()
 meth public abstract jakarta.data.Sort ascIgnoreCase()
 meth public abstract jakarta.data.Sort desc()
 meth public abstract jakarta.data.Sort descIgnoreCase()
 meth public abstract java.lang.String name()
 
-CLSS public abstract interface !annotation jakarta.data.model.StaticMetamodel
+CLSS public abstract interface !annotation jakarta.data.metamodel.StaticMetamodel
  anno 0 java.lang.annotation.Documented()
  anno 0 java.lang.annotation.Retention(java.lang.annotation.RetentionPolicy value=RUNTIME)
  anno 0 java.lang.annotation.Target(java.lang.annotation.ElementType[] value=[TYPE])
 intf java.lang.annotation.Annotation
 meth public abstract java.lang.Class<?> value()
 
-CLSS abstract interface jakarta.data.model.package-info
+CLSS abstract interface jakarta.data.metamodel.package-info
 
 CLSS abstract interface jakarta.data.package-info
